+++ inst/files/a.R
@@ -1,0 +1,9 @@
+foo <- function(x) {
+    # print("this is a lint")
+    return(NULL)
+}
+
+bar <- function(x) {
+    # this is not a lint
+    return(NULL)
+}