+++ inst/files/a_verbatim.R
@@ -1,0 +1,10 @@
+foobar <- function(x) {
+    # this is a lint: this line is way toooooooooooooooooooooooooooooooooooo long!
+    return(NULL)
+}
+
+barfoo <- function(x) {
+    # this is not a lint
+    return(NULL)
+}
+